+++ model/src/main/java/edu/java/exceptions/BadRequestException/ChatAlreadyRegisteredException.java
@@ -3,23 +3,23 @@ package edu.java.exceptions.BadRequestException;
 import java.util.List;
 
 public class ChatAlreadyRegisteredException extends BadRequestException {
-    public ChatAlreadyRegisteredException(List<Integer> ids, String uri, String message) {
+    public ChatAlreadyRegisteredException(List<Long> ids, String uri, String message) {
         super(ids, uri, message);
     }
 
-    public ChatAlreadyRegisteredException(List<Integer> ids, String uri) {
+    public ChatAlreadyRegisteredException(List<Long> ids, String uri) {
         this(ids, uri, DEFAULT_MESSAGE);
     }
 
-    public ChatAlreadyRegisteredException(Integer id, String uri) {
+    public ChatAlreadyRegisteredException(long id, String uri) {
         this(List.of(id), uri);
     }
 
-    public ChatAlreadyRegisteredException(Integer id) {
+    public ChatAlreadyRegisteredException(long id) {
         this(id, "");
     }
 
-    public Integer id() {
+    public long id() {
         return this.ids().getFirst();
     }
 }
