+++ model/src/main/java/edu/java/exceptions/BadRequestException/ChatsNotRegisteredException.java
@@ -3,11 +3,11 @@ package edu.java.exceptions.BadRequestException;
 import java.util.List;
 
 public class ChatsNotRegisteredException extends BadRequestException {
-    public ChatsNotRegisteredException(List<Integer> ids, String uri, String message) {
+    public ChatsNotRegisteredException(List<Long> ids, String uri, String message) {
         super(ids, uri, message);
     }
 
-    public ChatsNotRegisteredException(List<Integer> ids, String uri) {
+    public ChatsNotRegisteredException(List<Long> ids, String uri) {
         this(ids, uri, DEFAULT_MESSAGE);
     }
 }
