+++ model/src/main/java/edu/java/exceptions/NotFoundException/ChatNotTrackedUriException.java
@@ -1,11 +1,11 @@
 package edu.java.exceptions.NotFoundException;
 
 public class ChatNotTrackedUriException extends NotFoundException {
-    public ChatNotTrackedUriException(Integer id, String uri, String message) {
+    public ChatNotTrackedUriException(long id, String uri, String message) {
         super(id, uri, message);
     }
 
-    public ChatNotTrackedUriException(Integer id, String uri) {
+    public ChatNotTrackedUriException(long id, String uri) {
         this(id, uri, DEFAULT_MESSAGE);
     }
 }
