+++ model/src/main/java/edu/java/exceptions/NotFoundException/NotFoundException.java
@@ -2,16 +2,16 @@ package edu.java.exceptions.NotFoundException;
 
 public abstract class NotFoundException extends Exception {
     protected static final String DEFAULT_MESSAGE = "Not found";
-    protected final Integer id;
+    protected final long id;
     protected final String uri;
 
-    protected NotFoundException(Integer id, String uri, String message) {
+    protected NotFoundException(long id, String uri, String message) {
         super(message);
         this.id = id;
         this.uri = uri;
     }
 
-    public Integer id() {
+    public long id() {
         return id;
     }
